Fixed Tonnage Issue by changing to double with bigger limit

+++ app/database/migrations/2014_04_05_144219_create_fleets.php
@@ -20,7 +20,7 @@ class CreateFleets extends Migration {
 			$table->string('empire');
 			$table->string('faction');
 			$table->integer('ships');
-			$table->float('tonnage');
+			$table->double('tonnage',13,1);
 			$table->integer('warfacts_id')->nullable();
 			$table->integer('x');
 			$table->integer('y');
